server/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
attrs==25.3.0
certifi==2025.6.15
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
ecdsa==0.19.1
fastapi==0.115.14
greenlet==3.2.3
h11==0.16.0
httptools==0.6.4
idna==3.10
orjson==3.10.18
outcome==1.3.0.post0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
PySocks==1.7.1
python-dotenv==1.1.1
python-jose==3.5.0
PyYAML==6.0.2
rsa==4.9.1
selenium==4.34.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
starlette==0.46.2
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.35.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
